requirements.txt
comtypes==1.1.4
imutils==0.4.5
numpy==1.14.0
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
pykinect2==0.1.0
Pillow==5.0.0
